# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'New India Export / VIRASTRA API'
  version: 2.3.0
  description: "Production Express API for customer onboarding, OTP auth, KYC, cases, Razorpay payments, GST invoices, events, and admin ops.\n\n**Auth:** Email OTP only. Send `Authorization: Bearer <JWT>` for protected routes.\n\n**Money:** Amounts in INR rupees in API/DB; converted to paise only for Razorpay.\n\n**Envelope:** `{ success, data?, message? }` on success; `{ success: false, message, code?, details? }` on errors.\n\nLegacy vault/shipment routes under `/api/cases/...` exist only when `ENABLE_LEGACY_SHIPMENT=true`."
  contact:
    name: 'New India Export'
    email: support@newindiaexport.com
servers:
  -
    url: 'http://127.0.0.1:5001'
    description: Local
  -
    url: /
    description: 'Current host'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from OTP verify / login. Claims: sub, email, role, status.'
...
  schemas:
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        message:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        code:
          type: string
        details:
          type: array
          items: {}
...
    StaffAccessPatch:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - Pending
            - Approved
            - Rejected
            - Suspended
        permissions:
          type: array
          items:
            type: string
        reason:
          type: string
    StaffUserPatch:
      type: object
      properties:
        role:
          type: string
          enum:
            - customer
            - operations
            - admin
        status:
          type: string
        name:
          type: string
        permissions:
          type: array
          items:
            type: string
    StaffPermissions:
      type: object
      required:
        - permissions
      properties:
        permissions:
          type: array
          items:
            type: string
...
    SupportTicket:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        message:
          type: string
    LeadContact:
      type: object
      required:
        - email
        - name
        - message
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        company:
          type: string
        message:
          type: string
        source:
          type: string
    ConciergeBook:
      type: object
      required:
        - email
        - name
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        phone:
          type: string
        company:
          type: string
        preferredSlot:
          type: string
        message:
          type: string
...
  responses:
    BadRequest:
      description: 'Validation / bad request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 'Missing or invalid JWT'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 'Insufficient role / ACL'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 'Resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/staff/access-requests:
    get:
      summary: 'List staff access requests'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: status
          in: query
          required: false
          schema:
            type: string
        -
          name: q
          in: query
          required: false
          schema:
            type: string
  '/api/staff/access-requests/{reqId}':
    get:
      summary: 'Get staff access request'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: reqId
          in: path
          required: true
          schema:
            type: string
          description: null
    patch:
      summary: 'Approve / reject / suspend request'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: reqId
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffAccessPatch'
  /api/staff/users:
    get:
      summary: 'List staff users'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  '/api/staff/users/{email}':
    patch:
      summary: 'Update staff user'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: email
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUserPatch'
  '/api/staff/users/{email}/permissions':
    patch:
      summary: 'Set staff permissions'
      tags:
        - Staff
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: email
          in: path
          required: true
          schema:
            type: string
          description: null
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffPermissions'
...
  /api/support/faqs:
    get:
      summary: FAQs
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
  /api/support/articles:
    get:
      summary: 'Help articles'
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
  /api/support/tickets:
    post:
      summary: 'Create support ticket'
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportTicket'
    get:
      summary: 'List support tickets (ops/admin)'
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security:
        -
          bearerAuth: []
  /api/leads/contact:
    post:
      summary: 'Contact lead'
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LeadContact'
  /api/support/concierge/book:
    post:
      summary: 'Book concierge call'
      tags:
        - Support
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConciergeBook'
...